+++ gimie/sources/github.py
@@ -18,7 +18,6 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from datetime import datetime
-import requests
 import os
 from typing import Any, Dict, List, Optional, Set, Union
 from urllib.parse import urlparse
@@ -37,41 +36,11 @@ from gimie.models import (
     IRI,
 )
 from gimie.graph.namespaces import SDO
-from gimie.utils import get_spdx_url
+from gimie.helpers_license import get_spdx_url
+from gimie.helpers_queries import send_rest_query, query_graphql,send_graphql_query
 
 GH_API = "https://api.github.com"
 load_dotenv()
-
-
-def send_rest_query(query: str, headers: Dict[str, str]) -> Dict[str, Any]:
-    """Generic function to send a GraphQL query to the GitHub API."""
-    resp = requests.get(
-        url=f"{GH_API}/{query}",
-        headers=headers,
-    )
-
-    if resp.status_code != 200:
-        raise ConnectionError(resp.json()["message"])
-    return resp.json()
-
-
-def send_graphql_query(
-    query: str, data: Dict[str, Any], headers: Dict[str, str]
-) -> Dict[str, Any]:
-    """Generic function to send a GraphQL query to the GitHub API."""
-    resp = requests.post(
-        url=f"{GH_API}/graphql",
-        json={
-            "query": query,
-            "variables": data,
-        },
-        headers=headers,
-    )
-
-    if resp.status_code != 200:
-        raise ConnectionError(resp.json()["message"])
-    return resp.json()
-
 
 def query_contributors(
     url: str, headers: Dict[str, str]
@@ -82,7 +51,7 @@ def query_contributors(
     owner, name = urlparse(url).path.strip("/").split("/")
     # Get contributors (available in the REST API but not GraphQL)
     contributors = send_rest_query(
-        f"repos/{owner}/{name}/contributors", headers=headers
+        api = GH_API, f"repos/{owner}/{name}/contributors", headers=headers
     )
     ids = [contributor["node_id"] for contributor in contributors]
     # Get all contributors' metadata in 1 GraphQL query
@@ -109,93 +78,10 @@ def query_contributors(
     }"""
 
     contributors = send_graphql_query(
-        users_query, data={"ids": ids}, headers=headers
+        api = GH_API, users_query, data={"ids": ids}, headers=headers
     )
     # Drop empty users (e.g. dependabot)
     return [user for user in contributors["data"]["nodes"] if user]
-
-
-def query_repo_graphql(url: str, headers: Dict[str, str]) -> Dict[str, Any]:
-    """Queries the GitHub GraphQL API to extract metadata about
-    target repository.
-
-    Parameters
-    ----------
-    url:
-        URL of the repository to query.
-    """
-    owner, name = urlparse(url).path.strip("/").split("/")
-    repo_query = """
-    query repo($owner: String!, $name: String!) {
-        repository(name: $name, owner: $owner) {
-            createdAt
-            description
-            latestRelease {
-                name
-            }
-            licenseInfo {
-                spdxId
-            }
-            mentionableUsers(first: 100) {
-                nodes {
-                    login
-                    name
-                    avatarUrl
-                    company
-                    organizations(first: 100) {
-                        nodes {
-                            avatarUrl
-                            description
-                            login
-                            name
-                            url
-                        }
-                    }
-                    url
-                }
-            }
-            name
-            owner {
-                avatarUrl
-                login
-                url
-                ... on User {
-                    company
-                    name
-                    organizations(first: 100) {
-                        nodes {
-                            avatarUrl
-                            description
-                            login
-                            name
-                            url
-                        }
-                    }
-                }
-                ... on Organization {
-                    name
-                    description
-                }
-            }
-            primaryLanguage {
-                name
-            }
-            repositoryTopics(first: 10) {
-                nodes {
-                    topic {
-                        name
-                    }
-                }
-            }
-            updatedAt
-            url
-        }
-    }
-    """
-    repo = send_graphql_query(
-        repo_query, data={"owner": owner, "name": name}, headers=headers
-    )
-    return repo
 
 
 @dataclass
@@ -204,7 +90,7 @@ class GithubExtractor(Extractor):
     extract metadata into linked data."""
 
     path: str
-    github_token: Optional[str] = None
+    token: Optional[str] = None
 
     name: Optional[str] = None
     author: Optional[Union[Organization, Person]] = None
@@ -250,7 +136,76 @@ class GithubExtractor(Extractor):
 
     def _fetch_repo_data(self, url: str) -> Dict[str, Any]:
         """Fetch repository metadata from GraphQL endpoint."""
-        response = query_repo_graphql(url, self._set_auth())
+        owner, name = urlparse(url).path.strip("/").split("/")
+        data = {"owner": owner, "name": name}
+        repo_query = """
+        query repo($owner: String!, $name: String!) {
+            repository(name: $name, owner: $owner) {
+                createdAt
+                description
+                latestRelease {
+                    name
+                }
+                licenseInfo {
+                    spdxId
+                }
+                mentionableUsers(first: 100) {
+                    nodes {
+                        login
+                        name
+                        avatarUrl
+                        company
+                        organizations(first: 100) {
+                            nodes {
+                                avatarUrl
+                                description
+                                login
+                                name
+                                url
+                            }
+                        }
+                        url
+                    }
+                }
+                name
+                owner {
+                    avatarUrl
+                    login
+                    url
+                    ... on User {
+                        company
+                        name
+                        organizations(first: 100) {
+                            nodes {
+                                avatarUrl
+                                description
+                                login
+                                name
+                                url
+                            }
+                        }
+                    }
+                    ... on Organization {
+                        name
+                        description
+                    }
+                }
+                primaryLanguage {
+                    name
+                }
+                repositoryTopics(first: 10) {
+                    nodes {
+                        topic {
+                            name
+                        }
+                    }
+                }
+                updatedAt
+                url
+            }
+        }
+        """
+        response = query_graphql(GH_API, repo_query, data, self._set_auth())
         if "errors" in response:
             raise ValueError(response["errors"])
 
@@ -269,10 +224,10 @@ class GithubExtractor(Extractor):
     def _set_auth(self) -> Any:
         """Set authentication headers for GitHub API requests."""
         try:
-            if not self.github_token:
-                self.github_token = os.environ.get("ACCESS_TOKEN")
-                assert self.github_token
-            headers = {"Authorization": f"token {self.github_token}"}
+            if not self.token:
+                self.token = os.environ.get("ACCESS_TOKEN")
+                assert self.token
+            headers = {"Authorization": f"token {self.token}"}
 
             login = requests.get(f"{GH_API}/user", headers=headers)
             assert login.json().get("login")
